Migrate CPShell to HH3 (v3) SDK
