+++ .devcontainer/devcontainer.json
@@ -1,0 +1,18 @@
+{
+    "build": {
+        "dockerfile": "Dockerfile"
+    },
+    "customizations": {
+        "vscode": {
+            "extensions": [
+                "ms-vscode.makefile-tools",
+                "ms-vscode.cpptools-themes",
+                "ms-vscode.cpptools",
+                "xaver.clang-format",
+                "llvm-vs-code-extensions.vscode-clangd",
+                "ZixuanWang.linkerscript",
+                "ms-azuretools.vscode-docker"
+            ]
+        }
+    }
+}
