--- linker.ld
@@ -1,5 +1,0 @@
-ENTRY(_main);
-
-SECTIONS {
-	. = 0x8CFF0000;
-}
